Remove sitePreview from site-react
