+++ packages/site/site-nextjs/src/blocks/DamVideoBlock.tsx
@@ -1,12 +1,10 @@
 "use client";
 
-import { useIsElementInViewport, withPreview } from "@comet/site-react";
-import { type PropsWithData } from "@comet/site-react";
+import { type PropsWithData, PreviewSkeleton, useIsElementInViewport, withPreview } from "@comet/site-react";
 import clsx from "clsx";
 import { type ReactElement, type ReactNode, useRef, useState } from "react";
 
 import { type DamVideoBlockData } from "../blocks.generated";
-import { PreviewSkeleton } from "../previewskeleton/PreviewSkeleton";
 import styles from "./DamVideoBlock.module.scss";
 import { type VideoPreviewImageProps, VideoPreviewImage } from "./helpers/VideoPreviewImage";
