+++ packages/site/site-nextjs/src/blocks/PixelImageBlock.tsx
@@ -6,13 +6,13 @@ import {
     generateImageUrl,
     getMaxDimensionsFromArea,
     parseAspectRatio,
+    PreviewSkeleton,
     withPreview,
 } from "@comet/site-react";
 // eslint-disable-next-line no-restricted-imports
 import NextImage, { type ImageProps } from "next/image";
 
 import { type PixelImageBlockData } from "../blocks.generated";
-import { PreviewSkeleton } from "../previewskeleton/PreviewSkeleton";
 import styles from "./PixelImageBlock.module.scss";
 
 interface PixelImageBlockProps extends PropsWithData<PixelImageBlockData>, Omit<ImageProps, "src" | "width" | "height" | "alt"> {
