+++ packages/site/site-nextjs/src/blocks/VimeoVideoBlock.tsx
@@ -1,11 +1,9 @@
 "use client";
-import { useIsElementInViewport, withPreview } from "@comet/site-react";
-import { type PropsWithData } from "@comet/site-react";
+import { type PropsWithData, PreviewSkeleton, useIsElementInViewport, withPreview } from "@comet/site-react";
 import clsx from "clsx";
 import { type ReactNode, useRef, useState } from "react";
 
 import { type VimeoVideoBlockData } from "../blocks.generated";
-import { PreviewSkeleton } from "../previewskeleton/PreviewSkeleton";
 import { type VideoPreviewImageProps, VideoPreviewImage } from "./helpers/VideoPreviewImage";
 import styles from "./VimeoVideoBlock.module.scss";
