+++ packages/site/site-nextjs/src/index.ts
@@ -18,27 +18,20 @@ export {
     createGraphQLFetch,
     gql,
 } from "./graphQLFetch/graphQLFetch";
-// export { IFrameBridgeProvider } from "./iframebridge/IFrameBridge";
-// export { IFrameMessageType } from "./iframebridge/IFrameMessage";
-// export { Preview } from "./iframebridge/Preview";
 export { useBlockPreviewFetch } from "./iframebridge/useBlockPreviewFetch";
-// export { useIFrameBridge } from "./iframebridge/useIFrameBridge";
-// export { type WithPreviewProps, isWithPreviewPropsData, withPreview } from "./iframebridge/withPreview";
 export { Image } from "./image/Image";
-export { PreviewSkeleton } from "./previewskeleton/PreviewSkeleton";
 export { previewParams, sitePreviewRoute } from "./sitePreview/appRouter/sitePreviewRoute";
 export { sendSitePreviewIFrameMessage } from "./sitePreview/iframebridge/sendSitePreviewIFrameMessage";
 export { SitePreviewIFrameMessageType } from "./sitePreview/iframebridge/SitePreviewIFrameMessage";
 export { legacyPagesRouterSitePreviewApiHandler } from "./sitePreview/pagesRouter/legacyPagesRouterSitePreviewApiHandler";
 export { SitePreviewProvider } from "./sitePreview/SitePreviewProvider";
+export { type SitePreviewData, type SitePreviewParams } from "./sitePreview/SitePreviewUtils";
 export {
     type BlockLoader,
     type BlockLoaderDependencies,
     type CookieApi,
     type ImageDimensions,
     type PropsWithData,
-    type SitePreviewData,
-    type SitePreviewParams,
     type SupportedBlocks,
     type WithPreviewProps,
     BlockPreviewProvider,
@@ -59,6 +52,7 @@ export {
     OptionalBlock,
     parseAspectRatio,
     Preview,
+    PreviewSkeleton,
     recursivelyLoadBlockData,
     SvgImageBlock,
     useCookieApi,
