+++ packages/site/site-nextjs/src/sitePreview/SitePreviewUtils.ts
@@ -1,8 +1,16 @@
-import { type SitePreviewParams } from "@comet/site-react";
 import { errors, jwtVerify } from "jose";
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 export type Scope = Record<string, any>;
+
+export type SitePreviewData = {
+    includeInvisible: boolean;
+};
+export type SitePreviewParams = {
+    scope: Scope;
+    path: string;
+    previewData?: SitePreviewData;
+};
 
 export async function verifySitePreviewJwt(jwt: string): Promise<SitePreviewParams | null> {
     if (!process.env.SITE_PREVIEW_SECRET) {
