+++ packages/site/site-nextjs/src/sitePreview/appRouter/sitePreviewRoute.ts
@@ -1,12 +1,11 @@
 import "server-only";
 
-import { type SitePreviewParams } from "@comet/site-react";
 import { SignJWT } from "jose";
 import { cookies, draftMode } from "next/headers";
 import { redirect } from "next/navigation";
 import { type NextRequest, NextResponse } from "next/server";
 
-import { verifySitePreviewJwt } from "../SitePreviewUtils";
+import { type SitePreviewParams, verifySitePreviewJwt } from "../SitePreviewUtils";
 
 export async function sitePreviewRoute(request: NextRequest, _graphQLFetch: unknown /* deprecated: remove argument in v8 */) {
     const params = request.nextUrl.searchParams;
