--- packages/site/site-react/src/sitePreview/iframebridge/sendSitePreviewIFrameMessage.ts
@@ -1,5 +1,0 @@
-import { type SitePreviewIFrameMessage } from "./SitePreviewIFrameMessage";
-
-export function sendSitePreviewIFrameMessage(message: SitePreviewIFrameMessage) {
-    window.parent.postMessage(JSON.stringify(message), "*");
-}
